feat: refactor HTML rendering into handlebars templates
