+++ package.json
@@ -1,14 +1,24 @@
 {
   "name": "events-api-tampa-dev",
   "version": "0.0.0",
+  "type": "module",
+  "dependencies": {
+    "handlebars": "^4.7.7",
+    "hbs-async-render": "^1.0.1"
+  },
   "devDependencies": {
     "@cloudflare/workers-types": "^4.20230518.0",
+    "hbs-import-transpile": "^1.0.4",
+    "mocha": "^10.2.0",
     "typescript": "^5.1.3",
-    "wrangler": "2.1.15"
+    "wrangler": "^3.37.0"
   },
   "private": true,
   "scripts": {
-    "start": "wrangler dev",
-    "deploy": "wrangler publish"
+    "start": "wrangler dev -l",
+    "deploy": "wrangler publish",
+    "handlebars": "npm run compilehbs && npm run transpilehbs && rm src/*-original.js",
+    "compilehbs": "handlebars -e hbs -f src/pages-original.js views/pages/ views/rss/ && handlebars -e hbs -p -f src/partials-original.js views/partials/",
+    "transpilehbs": "hbs-import-transpile src/pages-original.js > assets/pages.js && hbs-import-transpile src/partials-original.js > assets/partials.js"
   }
 }
