+++ views/pages/error.hbs
@@ -1,0 +1,15 @@
+<html>
+<head>
+<meta charset="utf-8">
+{{> analytics/widget this }}
+{{> style/widget this }}
+
+</head>
+<body>
+<h2 id="error-message-head">{{this.error_status}} - {{this.error_message}}</h2>
+
+<h5 id="error-message-description">Need help? Check out our <a href="https://github.com/TampaDevs/events.api.tampa.dev">documentation.</a></h5>
+
+<p id="powered-by"><a href="https://www.tampadevs.com/">Tampa Devs</a> Events API Widget - Powered by <a href="https://github.com/TampaDevs/events.api.tampa.dev">Cloudflare</a></p>
+</body>
+</html>
