+++ views/pages/events-html.hbs
@@ -1,0 +1,60 @@
+<html>
+
+<head>
+    {{> common/global this }}
+    {{> analytics/widget this }}
+    {{> style/widget this }}
+
+</head>
+
+<body class="bg-gray-50 text-gray-800 font-sans">
+    <div class="max-w-4xl mx-auto py-8 px-4">
+        <div class="text-center mb-12">
+            <h1 class="text-4xl font-semibold">Upcoming Events</h1>
+            <h4 class="mt-2 text-1xl font-semibold"><a class="text-blue-500 hover:text-blue-700" href="https://tampa.dev/">Back to All Meetups</a></h4>
+        </div>
+
+        {{#if events.length}}
+          <div class="space-y-8">
+          {{#each events}}
+                <div class="bg-white shadow rounded-lg p-6">
+                    <div class="w-11/12" id="{{ this.group }}">
+                        <h2 class="text-2xl mb-1 font-semibold">{{ this.groupName }} - {{ this.data.title }}</h2>
+                        <h3 class="text-lg mb-2 font-bold">{{ to_local_date this.data.dateTime }}</h3>
+                        <h4 class="text-xs w-1/3 font-bold">
+                        {{#if this.google_maps_url}}
+                          <a class="text-blue-500 hover:text-blue-700" href="{{ this.google_maps_url }}">{{ this.address }}</a>
+                        {{else}}
+                          {{ this.address }}
+                        {{/if}}</h4>
+                    </div>
+
+                    <div class="flex mt-8 justify-center items-center">
+                        <img class="rounded-md drop-shadow-sm" src="{{ this.data.imageUrl }}" alt="{{ this.data.title }}" />
+                    </div>
+
+                    <div class="flex mb-4 mt-6 justify-center items-center">
+                        <a href="{{ this.data.eventUrl }}">
+                            <button class="bg-transparent hover:bg-blue-500 text-blue-700 font-semibold hover:text-white py-2 px-4 border border-blue-500 hover:border-transparent rounded">
+                                RSVP
+                            </button>
+                        </a>
+                    </div>
+
+                    <div class="flex flex-col justify-center items-center">
+                        <p class="mt-4 mb-10 w-11/12">{{ render_long_description this.data.description }}</p>
+                        <a class="text-blue-500 hover:text-blue-700 text-xs" href="{{ this.data.eventUrl }}">View on Meetup.com</a>
+                    </div>
+                </div>
+            {{/each}}
+            </div>
+        {{else}}
+            <div class="bg-white shadow rounded-lg p-6">
+                <div class="flex mt-4 mb-4 justify-center items-center">
+                    <h2 class="text-2xl font-semibold">No upcoming events.</h2>
+                </div>
+            </div>
+        {{/if}}
+
+    </div>
+</body>
