+++ views/pages/widget-carousel.hbs
@@ -1,0 +1,37 @@
+<html>
+
+<head>
+    {{> common/global this }}
+    {{> analytics/widget this }}
+    {{> style/widget this }}
+
+</head>
+
+<body>
+    {{#if events.length}}
+        <div class="flex overflow-x-auto py-4 space-x-4">
+            {{#each events}}
+                <div class="shrink-0 bg-white rounded-xl shadow-md overflow-hidden w-full max-w-lg">
+                    <div class="flex h-full flex-col md:flex-row">
+                        <div class="md:shrink-0">
+                            <a href="{{ add_utm this.data.eventUrl }}" target="_blank"><img class="w-full md:w-48 object-cover h-full" src="{{ this.data.imageUrl }}" alt="{{ this.data.title }}"></a>
+                        </div>
+                        <div class="p-4 md:p-8">
+                            <div class="uppercase tracking-wide text-sm text-blue-500 font-semibold"><a href="{{ add_utm this.data.eventUrl }}" target="_blank">{{ to_local_date this.data.dateTime }}</a></div>
+                            <a href="{{ add_utm this.data.eventUrl }}" class="block mt-1 text-lg leading-tight font-medium text-black hover:underline" target="_blank">{{ this.groupName }} - {{ this.data.title }}</a>
+                            <p class="mt-2 text-slate-500"><a href="{{ add_utm this.data.eventUrl }}" target="_blank">{{ render_short_description this.data.description }}</a></p>
+                        </div>
+                    </div>
+                </div>
+            {{/each}}
+        </div>
+    {{else}}
+        <div class="bg-white shadow rounded-lg p-6">
+            <div class="flex mt-4 mb-4 justify-center items-center">
+                <h2 class="text-2xl font-semibold">No upcoming events.</h2>
+            </div>
+        </div>
+    {{/if}}
+</body>
+
+</html>
