+++ views/pages/widget-next-event.hbs
@@ -1,0 +1,35 @@
+<html>
+
+<head>
+    {{> common/global this }}
+    {{> analytics/widget this }}
+    {{> style/widget this }}
+    
+</head>
+
+<body>
+    {{#if events.length}}
+        {{#each events}}
+            <div class="max-w-md mx-auto bg-white rounded-xl shadow-md overflow-hidden md:max-w-2xl">
+                <div class="md:flex">
+                    <div class="md:shrink-0">
+                        <a href="{{ add_utm this.data.eventUrl }}" target="_blank"><img class="h-48 w-full object-cover md:h-full md:w-48" src="{{ this.data.imageUrl }}" alt="{{ this.data.title }}"></a>
+                    </div>
+                    <div class="p-8">
+                        <div class="uppercase tracking-wide text-sm text-blue-500 font-semibold"><a href="{{ add_utm this.data.eventUrl }}" target="_blank">{{ to_local_date this.data.dateTime }}</a></div>
+                        <a href="{{ add_utm this.data.eventUrl }}" class="block mt-1 text-lg leading-tight font-medium text-black hover:underline" target="_blank">{{ this.groupName }} - {{ this.data.title }}</a>
+                        <p class="mt-2 text-slate-500"><a href="{{ add_utm this.data.eventUrl }}" target="_blank">{{ render_short_description this.data.description }}</a></p>
+                    </div>
+                </div>
+            </div>
+        {{/each}}
+    {{else}}
+        <div class="bg-white shadow rounded-lg p-6">
+            <div class="flex mt-4 mb-4 justify-center items-center">
+                <h2 class="text-2xl font-semibold">No upcoming events.</h2>
+            </div>
+        </div>
+    {{/if}}
+</body>
+
+</html>
