+++ views/partials/style/widget.hbs
@@ -1,0 +1,1 @@
+<script src="https://cdn.tailwindcss.com"></script>
